Moved the matrix loading functions to the utility module and used it also in the creation of the coulomb matrix
