+++ src/quatrex/electron/solver.py
@@ -8,12 +8,8 @@ from qttools.comm import comm
 from qttools.datastructures import DSDBSparse
 from qttools.greens_function_solver.solver import OBCBlocks
 from qttools.profiling import Profiler, decorate_methods
-from qttools.utils.gpu_utils import get_host, synchronize_device
-from qttools.utils.input_utils import (
-    create_coordinate_grid,
-    create_hamiltonian,
-    cutoff_hr,
-)
+from qttools.utils.gpu_utils import synchronize_device
+from qttools.utils.input_utils import create_coordinate_grid
 from qttools.utils.mpi_utils import distributed_load, get_local_slice, get_section_sizes
 from qttools.utils.stack_utils import scale_stack
 
@@ -26,7 +22,13 @@ from quatrex.core.compute_config import ComputeConfig
 from quatrex.core.quatrex_config import QuatrexConfig
 from quatrex.core.statistics import fermi_dirac
 from quatrex.core.subsystem import SubsystemSolver
-from quatrex.core.utils import assemble_kpoint_dsb, get_periodic_superblocks, homogenize
+from quatrex.core.utils import (
+    assemble_kpoint_dsb,
+    get_periodic_superblocks,
+    homogenize,
+    load_matrix_from_files,
+    load_matrix_from_unit_cell,
+)
 
 profiler = Profiler()
 
@@ -136,6 +138,7 @@ class ElectronSolver(SubsystemSolver):
                 hamiltonian_dict,
                 number_of_kpoints,
                 0,
+                transport_direction=quatrex_config.device.transport_direction,
             )
         del hamiltonian_sparray
         del hamiltonian_dict
@@ -344,190 +347,6 @@ class ElectronSolver(SubsystemSolver):
             dtype=self.hamiltonian.dtype,
         )
 
-    def _load_matrix_from_unit_cell(
-        self, quatrex_config, matrix_name: str, use_r_cutoff: bool = True
-    ) -> tuple[sparse.coo_matrix, dict | None, NDArray | None]:
-        """Generic method to load a matrix from unit cell data.
-
-        Parameters
-        ----------
-        quatrex_config : QuatrexConfig
-            The quatrex simulation configuration.
-        matrix_name : str
-            Name of the matrix ('hamiltonian' or 'overlap').
-        use_r_cutoff : bool
-            Whether to apply R_cutoff to the unit cells.
-
-        Returns
-        -------
-        tuple[sparse.coo_matrix, dict | None, NDArray | None]
-            The matrix, optional k-point dictionary, and optional block sizes.
-        """
-        unit_cells = distributed_load(
-            quatrex_config.input_dir / f"{matrix_name}_unit_cells.npy"
-        ).astype(xp.complex128)
-
-        # Apply cutoff if requested and available
-        if use_r_cutoff and quatrex_config.device.R_cutoff is not None:
-            unit_cells = cutoff_hr(
-                unit_cells,
-                R_cutoff=quatrex_config.device.R_cutoff,
-            )
-        elif matrix_name == "overlap":
-            # For overlap, use unit_cell_per_supercell as R_cutoff
-            unit_cells = cutoff_hr(
-                unit_cells,
-                R_cutoff=quatrex_config.device.unit_cell_per_supercell,
-            )
-
-        return self._create_matrix_from_unit_cells(quatrex_config, unit_cells)
-
-    def _create_matrix_from_unit_cells(
-        self, quatrex_config, unit_cells
-    ) -> tuple[sparse.coo_matrix, dict | None, NDArray | None]:
-        """Generic method to create a matrix from unit cells with periodic shifts.
-
-        Parameters
-        ----------
-        quatrex_config : QuatrexConfig
-            The quatrex simulation configuration.
-        unit_cells : NDArray
-            The unit cell data.
-
-        Returns
-        -------
-        tuple[sparse.coo_matrix, dict | None, NDArray | None]
-            The matrix, optional k-point dictionary, and optional block sizes.
-        """
-        # Determine the local slice of the data.
-        # NOTE: This is arrow-wise partitioning.
-        # TODO: Allow more options, e.g., block row-wise partitioning.
-        section_sizes, __ = get_section_sizes(
-            quatrex_config.device.number_of_supercells, comm.block.size
-        )
-        section_offsets = np.hstack(([0], np.cumsum(section_sizes)))
-        start_block = section_offsets[comm.block.rank]
-        end_block = section_offsets[comm.block.rank + 1]
-
-        matrix_dict = {}
-        # Create the matrix for each periodic shift
-        for periodic_shift in xp.ndindex(
-            tuple(
-                2 * ps - 1 for ps in quatrex_config.device.cells_in_periodic_directions
-            )
-        ):
-            periodic_shift = tuple(
-                [
-                    ps - quatrex_config.device.cells_in_periodic_directions[i] + 1
-                    for i, ps in enumerate(periodic_shift)
-                ]
-            )
-            matrix_sparray, block_sizes = create_hamiltonian(
-                unit_cells,
-                quatrex_config.device.number_of_supercells,
-                quatrex_config.device.transport_direction,
-                quatrex_config.device.unit_cell_per_supercell,
-                block_start=start_block,
-                block_end=end_block,
-                periodic_shift=periodic_shift,
-                return_sparse=True,
-            )
-            matrix_dict[periodic_shift] = matrix_sparray.astype(xp.complex128)
-
-        matrix_sparray = sum(matrix_dict.values())
-        matrix_sparray.sum_duplicates()
-        block_sizes = get_host(block_sizes)
-        block_sizes_array = np.asarray(
-            [block_sizes[0]] * quatrex_config.device.number_of_supercells
-        )
-
-        return matrix_sparray, matrix_dict, block_sizes_array
-
-    def _load_matrix_from_files(
-        self, quatrex_config, matrix_name: str
-    ) -> tuple[sparse.coo_matrix, dict | None, NDArray | None]:
-        """Generic method to load a matrix from pre-computed files.
-
-        Parameters
-        ----------
-        quatrex_config : QuatrexConfig
-            The quatrex simulation configuration.
-        matrix_name : str
-            Name of the matrix ('hamiltonian' or 'overlap').
-
-        Returns
-        -------
-        tuple[sparse.coo_matrix, dict | None, NDArray | None]
-            The matrix, optional k-point dictionary, and optional block sizes.
-        """
-        # Define file loading priority for each matrix type
-        if matrix_name == "hamiltonian":
-            file_patterns = [
-                ("hamiltonian.npz", "npz"),
-                ("hamiltonian.pkl", "pkl"),
-            ]
-        else:  # overlap
-            file_patterns = [
-                ("overlap_matrix.pkl", "pkl"),
-                ("overlap.npz", "npz"),
-            ]
-
-        # Try loading files in priority order
-        for filename, file_type in file_patterns:
-            try:
-                if file_type == "pkl":
-                    matrix_dict = distributed_load(quatrex_config.input_dir / filename)
-                    matrix_sparray = sum(matrix_dict.values())
-                    matrix_sparray.sum_duplicates()
-                    block_sizes = self._load_block_sizes(quatrex_config, matrix_name)
-                    return matrix_sparray, matrix_dict, block_sizes
-
-                else:  # npz
-                    matrix_sparray = distributed_load(
-                        quatrex_config.input_dir / filename
-                    ).astype(xp.complex128)
-                    block_sizes = self._load_block_sizes(quatrex_config, matrix_name)
-                    return matrix_sparray, None, block_sizes
-
-            except FileNotFoundError:
-                continue
-
-        # If no files found, handle fallback
-        if matrix_name == "overlap":
-            # For overlap, fall back to identity matrix
-            return self._create_identity_overlap(), None, None
-        else:
-            # For Hamiltonian, this is an error
-            raise FileNotFoundError(
-                f"No {matrix_name} files found in {quatrex_config.input_dir}"
-            )
-
-    def _load_block_sizes(self, quatrex_config, matrix_name: str) -> NDArray | None:
-        """Load block sizes if available and needed.
-
-        Parameters
-        ----------
-        quatrex_config : QuatrexConfig
-            The quatrex simulation configuration.
-        matrix_name : str
-            Name of the matrix ('hamiltonian' or 'overlap').
-
-        Returns
-        -------
-        NDArray | None
-            Block sizes array or None if not needed/available.
-        """
-        if matrix_name == "hamiltonian":
-            try:
-                return get_host(
-                    distributed_load(quatrex_config.input_dir / "block_sizes.npy")
-                )
-            except FileNotFoundError:
-                raise FileNotFoundError(
-                    "block_sizes.npy required for Hamiltonian loading"
-                )
-        return None
-
     def _load_overlap_matrix(
         self, quatrex_config
     ) -> tuple[sparse.coo_matrix, dict | None]:
@@ -545,11 +364,11 @@ class ElectronSolver(SubsystemSolver):
         """
         try:
             if quatrex_config.device.construct_from_unit_cell:
-                matrix_sparray, matrix_dict, _ = self._load_matrix_from_unit_cell(
+                matrix_sparray, matrix_dict, _ = load_matrix_from_unit_cell(
                     quatrex_config, "overlap", use_r_cutoff=False
                 )
             else:
-                matrix_sparray, matrix_dict, _ = self._load_matrix_from_files(
+                matrix_sparray, matrix_dict, _ = load_matrix_from_files(
                     quatrex_config, "overlap"
                 )
             return matrix_sparray, matrix_dict
@@ -573,9 +392,9 @@ class ElectronSolver(SubsystemSolver):
             The Hamiltonian matrix, optional k-point dictionary, and block sizes.
         """
         if quatrex_config.device.construct_from_unit_cell:
-            return self._load_matrix_from_unit_cell(quatrex_config, "hamiltonian")
+            return load_matrix_from_unit_cell(quatrex_config, "hamiltonian")
         else:
-            return self._load_matrix_from_files(quatrex_config, "hamiltonian")
+            return load_matrix_from_files(quatrex_config, "hamiltonian")
 
     def update_potential(self, new_potential: NDArray) -> None:
         """Updates the potential matrix.
